version: 0.0.6
- add sider - testing collapsible
- add testing for sider
- add testing for menu
