+++ .husky/pre-commit
@@ -2,4 +2,5 @@
 . "$(dirname "$0")/_/husky.sh"
 
 npx lint-staged
+yarn test-antd
 yarn build
