+++ libs/antd/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@wsh4and/antd",
-  "version": "0.0.5",
+  "version": "0.0.6",
   "publishConfig": {
     "@wsh4and:registry": "https://registry.npmjs.org/"
   }
