--- libs/antd/src/lib/menu/menu.module.css
@@ -1,7 +1,0 @@
-/*
- * Replace this with your own classes
- *
- * e.g.
- * .container {
- * }
-*/
