+++ libs/antd/src/lib/sider/sider.tsx
@@ -1,17 +1,12 @@
 import { Layout, SiderProps } from 'antd';
 
 const { Sider } = Layout;
-
 export interface ISider extends SiderProps {
   children: React.ReactNode;
 }
 
 function NewSider(props: ISider) {
-  return (
-    <Sider {...props} width={200} className="site-layout-background">
-      {props.children}
-    </Sider>
-  );
+  return <Sider {...props}>{props.children}</Sider>;
 }
 
 export { NewSider as Sider };
